extract base64 routines
